Move enumerateParametersUsableInExpressions in its own file.
